feat: migrate gold price and tax to BigDecimal
- Migrated User entity fields to BigDecimal for precision.
- Added Room TypeConverters for BigDecimal support.
- Added unarchiveProduct functionality to ShopDao.
- Incremented database version to 9.

+++ app/src/test/java/com/example/domain/FinancialCalculationTest.kt
@@ -200,7 +200,7 @@ class FinancialCalculationTest {
         val discount = 500_000.0
 
         val itemsSumBd = items.fold(BigDecimal.ZERO) { acc, item ->
-            acc.add(BigDecimal.valueOf(item.total).setScale(0, RoundingMode.HALF_UP))
+            acc.add(item.total.setScale(0, RoundingMode.HALF_UP))
         }
         val discountBd = BigDecimal.valueOf(discount).setScale(0, RoundingMode.HALF_UP)
         val expectedPayable = itemsSumBd.subtract(discountBd).toDouble()
